+++ tests/_tplcache/demo_tpl__global.php
@@ -141,7 +141,7 @@ private function _blockfn_Block2(&$data, $self, $__i__)
         $__p__ .= '' . "\n" . '        unkonown result' . "\n" . '    ';
     }
     
-    $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '    ' . "\n" . '    <strong>SET a new tpl variable</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %set($foo, "123") %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ';
+    $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '    ' . "\n" . '    <strong>Inline (ternary) IF</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %iif( 1+1 == 2, "1+1 = 2", "1+1 = 1" ) %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ' . (Contemplate::iif( 1+1 == 2, "1+1 = 2", "1+1 = 1" )) . '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '    ' . "\n" . '    <strong>SET a new tpl variable</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %set($foo, "123") %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ';
     $data['foo'] = ("123");
     
     $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '    ' . "\n" . '    <strong>CHECK ISSET for a tpl variable</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %if( %isset($foo) ) %&gt;' . "\n" . '        $foo is SET' . "\n" . '    &lt;% %else() %&gt;' . "\n" . '        $foo is NOT SET' . "\n" . '    &lt;% %fi() %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ';
@@ -156,13 +156,25 @@ private function _blockfn_Block2(&$data, $self, $__i__)
         $__p__ .= '' . "\n" . '        $foo is NOT SET' . "\n" . '    ';
     }
     
-    $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '' . "\n" . '    <!-- include a (sub-)template file -->' . "\n" . '    <strong>INCLUDE a (sub-)template file</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %include("date") %&gt;' . "\n" . '    </pre><br />' . "\n" . '     <!-- print a localized date php-style -->' . "\n" . '<strong>A (localized) date, PHP-style</strong><br />' . "\n" . '' . (Contemplate::ldate("M, d")) . '' . "\n" . '';
+    $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '' . "\n" . '    <strong>CHECK EMPTY for a tpl variable</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %if( %empty($foo) ) %&gt;' . "\n" . '        $foo is EMPTY' . "\n" . '    &lt;% %else() %&gt;' . "\n" . '        $foo is NOT EMPTY' . "\n" . '    &lt;% %fi() %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ';
+    if ( empty($data['foo']) )
+    {
+        
+        $__p__ .= '' . "\n" . '        $foo is EMPTY' . "\n" . '    ';
+    }
+    else
+    {
+        
+        $__p__ .= '' . "\n" . '        $foo is NOT EMPTY' . "\n" . '    ';
+    }
+    
+    $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '' . "\n" . '    <!-- include a (sub-)template file -->' . "\n" . '    <strong>INCLUDE a (sub-)template file</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %include("date") %&gt;' . "\n" . '    </pre><br />' . "\n" . '     <!-- print a localized date php-style -->' . "\n" . '<strong>A (localized) date, PHP-style</strong><br />' . "\n" . '' . (Contemplate::ldate("M, D, d")) . '' . "\n" . '';
     
     $__p__ .= '' . "\n" . '    ' . "\n" . '    <br /><br />' . "\n" . '' . "\n" . '    <strong>CALL another (sub-)template</strong><br />' . "\n" . '    <pre>' . "\n" . '    &lt;% %for($users as $i=>$usergroup) %&gt;' . "\n" . '        &lt;!-- call a (sub-)template --&gt;' . "\n" . '        &lt;% %tpl("sub", {"i" : $i, "users" : $users}) %&gt;' . "\n" . '    &lt;% %endfor() %&gt;' . "\n" . '    </pre><br />' . "\n" . '    ';
-    $_loc_81 = $data['users'];
-    if (!empty($_loc_81))
+    $_loc_84 = $data['users'];
+    if (!empty($_loc_84))
     {
-        foreach ($_loc_81 as $_loc_i=>$_loc_usergroup)
+        foreach ($_loc_84 as $_loc_i=>$_loc_usergroup)
         {
             
             $__p__ .= '' . "\n" . '        <!-- call a (sub-)template -->' . "\n" . '        ' . (Contemplate::tpl("sub", array("i" => $_loc_i, "users" => $data['users']))) . '' . "\n" . '    ';